requirements.txt
pytest
dragonfly2
pynput
python-docx
pycaw
mss